タグごとの計算方法をuseGraphData.tsでコンポーネント化
